updates should now work without separate server
this commit includes an extension that hosts a nuts-server to provide update information. This is only a temporary measure during development.
Also some cleanup after previous experiments.

And added another function to the extension context: "once". This can be called to specify a callback that will be called once, after all extensions are loaded and the redux store is initialized but before any other initialization happens.