Downloads are again global
This means that the store always contains the list of all downloads but they are still kept in different directories based on game.
The gameid is also stored inside the store so that we can filter/sort based on the game.

Having all downloads global has the advantage that we don't have to stop downloads when changing game modes.
Also, this gives more power to potential extensions like "clean up downloads older than x"