[#139948069] implemented script extender delegates
this also changes the core delegate object to receive the game id as a parameter instead of assuming the active game. The idea is that since the download view covers all games, if one installs an archive associated with Skyrim, it should always be installed for skyrim, not for the game currently being managed.