[#134914123] initial commit of local dependencies
local dependencies are stored with a mod instance installed on the current system and managed by the user, whereas "normal" rules are assigned to a mod through matching rules and are provided by a server.

This mainly adds an extension that shall eventually allow convenient management of these rules.
There are also a couple of  bugfixes and enhancements to how NMM2 integrates info from the dependency db into the manager.