bugfix: loadorder updated when switching to a game without plugins
previously the gamebryo-plugin extension was completely inactive when switching to an unsupported game, leaving the load order of the old game in the state.
This had the problem that checks and state-change handlers watching the loadorder didn't fire and parts of the application could still show info about those plugins.