[#130296011] simple ui to enable/disable extensions
- no install/update capabilities yet
- this actually required a fairly big change in how the store is initialised on startup because previously we loaded the store settings after extensions were loaded since extensions extend the store.
Now however we need to load (part of) the store before we can load extensions because how would the state control which extensions get loaded?

The way this is solved is that the store is loaded twice, first without extensions and only a single section "app". Once extensions are loaded, the store is completely recreated (using the same persistence backend).

In the process there was another change: Since we only load part of the store at first, we can't allow redux-persist to remove unknown keys from the persisted data. This change makes sense anyway because we don't want to drop data from an extension just because it has trouble loading once.