[#145268889] added an instance id
the instance id is a unique id generated on startup if none is set yet.
This id can later be used to differentiate between vortex instances on the same system or to identify a user when we can't or don't want to use private info like his user account.
The id is completely random (uuid v4) so it doesn't store any system or user identification.