ui polish
a large part of this commit is a new script and process and import svg icons from nucleo.
This now lets us pick the fonts we want to use from the nucleo app, transform them a bit (the nucleo app doesn't allow font rename) and then import them into the page as one file.
This should also be faster and shrink the dom considerably compared to before.