fixed crash on restart
this is more a workaround than anything else. Due to a bug in current version of electron (should be fixed in 1.7.7) the old Vortex instance stays arount a few seconds after shutdown which then keeps the new instance from accessing the state database.
This fix makes Vortex wait for the db to be unlocked but this will still cause a few seconds wait until the electron bug is fixed.