substantial redesign of stop-folder management
even though this is now implemented different than in NMM it fixes a few incompatible modules where the virtual install was actually incorrect but during "activation" the links were put in the right places. Since we want all virtual installations to be mapped 1:1 to the game directory and keep the hacks in the installer localised, stop folder management has to get everything right.