added support for reading the .ccc files
these files were introduced in fallout4 and skyrim se to list the creation club content so they don't have to be hard-coded in the exe (and thereby in every tool).
We still have a hardcoded list of the content that exists at this point to support the versions before the .ccc was introduced.