rewrote loot bindings
the native part of the loot lib is now synchronous.
To achive asynchronous behaviour (which is now optional) the js part of the lib spawns a separate process that does all the loot operations.
This is quite a bit more fiddly but should also be more robust, even if loot causes a crash it would be contained to that process.
Plus the loot operations now have no impact on the ui whatsoever.