added a builder config for a portable install
The portable won't be used for the moment because it's a bit pointless really.
What it is is an installer that is configured to extract vortex to a temporary location and run it from there.
I don't see the advantage over a one-click installer and it makes the install location intransparent to the user.