fixed: registered globally as handler for http protocol
was fortunately not _that_ big a deal because applications usually look for this registration, the sims 3 launcher is the only application I know of that actually used this.
this actually also fixes the problem that deregistering a protocol
didn't work which again isn't a big deal because it doesn't matter much whether
there is _no_ protocol handler or a non-working one.