nexus info is now stored persistently
previously vortex would always have to contact nexus to determine if a user is premium, causing the "go premium" banners to show up for a moment on startup before they disappear.
Now we assume the premium/supporter state is the same as last time until the update has been retrieved.