avoid lazy imports and handle (some) failed lazy imports
There is an odd, very rare, bug where Vortex fails to import modules that are part of the application bundle (the asar archive).
This seems to happen only if the module is being loaded after the initial startup, so a workaround is avoiding delayed loading.
Obviously this will slightly increase startup time which is also not great...