fixed: exception during purge if a directory is changed by an external application
the error is extremely rare: At the time the purge comes across the directory it's empty (apart from the __delete_if_empty tag) so Vortex wants to remove it, but when it does, the directory isn't empty any more. This is a race condition that is unlikely but we have to be prepared for it.
In another scenario something prevents vortex from removing the tag file but no error is produced. Not sure if/how this is possible but it would produce the same error

fixes #964