resolve custom protocols lazily and only to memory
the original url is stored with the download.
This means that nxm links for example are only resolved to real download urls when the download starts and if the download is paused, the client restarted, the link can be resolved again.

There are still severe problems with this commit, especially when it comes to setting the file name and progress of a resumed download.