warn user before dropping meta data about vanished mods
When the mod staging directory for a game is missing at startup Vortex has to drop meta information about the mods that were previously installed but now Vortex gives the user a chance to cancel out at that point. This may make sense for example if users have their mods installed on a removable drive and it's not connected (under the same drive letter).

If the user does decide to drop the mod meta information we now also (try to) purge files to ensure we have a clean slate