changed how table header are fixed
previously we used a css transform to keep the header visible.
This is unfortunately bad for performance because it prevents certain performance optimizations.
Now we have two tables, one that is just the header, with absolute positioning, and the other with the actual content.
The problem with this solution is that it usually requires fixed size columns as the header-only table doesn't have the content to correctly autosize to.
The fix for this is an update function that reads the width of the (hidden) header of the _real_ table and transfer those widths to the absolutely positioned one.