fixed: failure to install some mods
the problem occurred if the mod (c# fomod) installs directories in such a way that the output directory already exists. In this case if vortex tries to move the directory (as an optimization), the call would fail with an EPERM error.
It now falls back to copying in that case