fixed: inconsistent behaviour removing mods
between setting the status dropdown to uninstalled and using the remove button.
In the latter case we didn't refresh the mod conflict info.
In the former case, if you first disable the mod and then set it to uninstalled (without deploying, if manal deployment was enabled), Vortex didn't undeploy the mod and thus didn't clean up its files and instead trigger an external changes dialog