handle (theoretical?) issue if gamemode-activated emitted with wrong gameid
this is probably more of a precaution than a bugfix.
If this could happen, vortex would compare the mods in one game with those of another game and then logically find all of the mods in the one game "removed" and the one from the other one "added" which could seriously mess up the database.
I don't see how it could happen currently, right now it would require an error in the redux state watcher which would probably have produced a lot of very visible errors.