introduced a way to attach context info to errors
This is a bit hacky but may provide useful information for error searches.
The way it works is at the start of a long running process you can set context information for that process using "withContext()" which will automatically be unset when the process ends.
Should an error happen all current context information is also included in the report.
Since the context is a global it's completely possible that context is included that doesn't relate to the issue at all but that should be obvious from the error message/stack.
This way we should be able to see which mod was being installed when the installer fails, no matter where the error is thrown.