cleaned up "requiresCleanup" code
previously on removing a mod it would try to figure out which directories need to be cleaned up but the algorithm was O(n^2), potentially pretty slow (in case there were a lot of files added or removed) and in the "mergeMods === true" case it would return the base directory anyway.
Then it didn't handle errors.

Also, the cleanup this triggered completely ignored the "managed by vortex" tags and could therefore removed directories that were otherwise required.