fixed: multiple problems with symlink deployment
a) major: the non-elevated deployment method deleted the linked to file instead of the link during purge. The reason this is only noticed now is because in the past it was practically impossible to use that deployment method on windows, everyone was probably using the elevated variant. So this bug has only become serious with 0.17.10
b) a problem to remove a file during purge would cause a crash. Fortunately this is extremely rare (reported once in over a year)

fixes #4106