fixed: (rare) incorrect mod change detection
when switching to a game while installing a mod for that game, the mod refresh might detect the "installing" directory as a new mod and the actual mod id as deleted.

Not sure if this could actually cause any problem for the user since the mod in this state has no meta information anyway. It would only be reported since very recent releases