fallback for steam appId during launcher
Using the steamAppId game extensions set in "details" as a fallback if the game extension requests to be launched through steam.
This isn't ideal if the game extension can be used with multiple variants of the game but the game path matching is unreliable and this should improve the success rate considerably.

fixes #3889