fixed: runExecutable returned immediately when elevating
To clarify: runExecutable usually should return after the process has finished. When the executable has to be run elevated though it would return immediately after the process has been spawned.

This could lead to problems for example when starting fnis automatically as part of the deployment process because it would try to deploy files when the animations haven't been written yet