allow users to install non-archive files as mods
Vortex will offer to turn that file into a single-file mod

This is a bit of a hack because we always try to extract the file as an archive to support extensions like .fomod that are not archive extensions but are still archives 7zip supports.
That means we first provoke an error and then have an error handler that looks at the file extension and then makes the offer.