new function to purge a custom directory
Meaning we can purge a directory that is not/no longer the deployment target Vortex would use to deploy.
This is useful when an update has changed the deployment target and we need to clean up the legacy location.
Since Vortex only purges files that are recognized as links to the staging folder this should be perfectly safe.