fixed: update information dropped on network error
we drop information about mod updates if the mod is no longer available on nexus mods since you wouldn't be able to fetch it anyway.
But this also dropped the information if there was just a (temporary) network error.
Combined with the optimized update check this could lead to a state where the optimized check didn't find the update again later