added support for different extension types
the goal here is to allow translations and themes to be installed through the extension system and also for translations and games to show up in their respective pages to easier find them.

We could do the same fro themes at some point but right now we don't have a way to preview themes on the settings screen so having the screenshots on the extensions page is pretty essential to the user anyway.