fixed: dropdown menu overlapping custom titlebar broken
the way browsers/chrome implement the drag-area ignores z-layering of dom elements and thus the default is that the entire top area will be used for dragging the windows unless we specifically override that per css-class.
Apparently a recent update caused a change to the css class of the dropdown menu in the top right and removed the rule that fixed this.

fixes #5451