fixed crash on profiling
this was apparently caused by a bug in electron regarding wasm modules, the "fix" is merely that we don't load msgpack (our only wasm module at this time) until it's required.

Also this adds profiling to the menu for quicker access (generates more data than the inspector though). Hotkey is ctrl+shift+p