Added Browser Assistant specific installation error message
Browser assistant is one of two things:
- A tool developed and maintained by "Glarysoft"
- A Trojan developed by "Realistic Media Inc."
Regardless of who developed it, it shouldn't insert itself globally in the
manner that it seems to do, and Vortex definitely cannot function correctly
while it is installed.

There's not much we can do on Vortex's side of things besides warn the user
that he has this application installed, and to remove it if they wish to
use Vortex.

P.S. "Browser Assistant" related errors are no longer reportable.