fixed: mod references in rules not updated upon reinstall
that is: for the purpose for quicker and reliable lookup we stored the mod id in the rule once a mod was installed to satisfy the rule.
Afterwards however we would never remove/update the id so if the referenced mod got removed and a newer version installed to satisfy it, the rule may no longer be marked as fulfilled correctly.