cleand up full game search ui
The paths to search are no longer configurable through settings, Vortex will ask before every scan.

The reason for this is that waaaay too many users complain about Vortex not finding their game, presumably because Vortex wasn't searching the corresponding drive in the first place and the user never bothered to check the setting.

Since game serach isn't done a lot (0-1 times per users I'm going to presume) this doesn't actually make it more inconvenient, it merely ensures the user knows which drives are being searched.

This also improves the progress bar so it doesn't reach 100% before the search is done.