chrome fix now gets rid of manual dl confirmation
In the past the fix was intended to fix a situation where the nxm protocol got rejected entirely, the fix restored functionality.
The default, if nxm was not in excluded_schemes was that nxm would work so our fix didn't do anything.
Since chrome 77 the default (if a protocol is not listed) is to ask the user for confirmation and the clear cache function will reset all protocols to that default.
The fix in the original purpose is therefore not really necessary any more but instead it now users to get rid of the confirmation modal.