don't crash if state db contains key that don't have data
at least that's how I interpret the error happening. Essentially the list of keys returns a key that, upon "get"ing it causes a NotFoundError instead of - say - returning null or undefined.
This is extremely rare and we got no feedback as to what the user was doing so this shouldn't affect most users.