fixed: files from foreign mods removed during mod disable
when deactivating a mod Vortex would remove all files from the game folder that were contained in that mod, even when some of those files were actually deployed from a different mod.
This isn't a big problem though as a new deployment (which was required) would restore those files.
This change makes it so those files aren't removed and technically we probably wouldn't need a deployment any more.