fallback for cases where initial redux state can't be ipced
Not sure yet why this happens in the first place, it could simply be memory corruption on the user system. We now repeat, transfering the data chunked and in msgpack format.
This should deal with the problem if it's a temporary issue (memory corruption), a problem with too large ipcs or with JSON serialization.

Problem with this is that msgpack serializes "undefined" values different than json so in cases where this fallback is activated there may be new issues with code that assumes unset state is always "undefined"