force UAC consent window to the front
at least for the creation of symbolic links we can now try to force the UAC window to be on top. I wasn't able to reproduce it but based on the reports of some users it opened behind the Vortex.

This will also cancel the deployment if the consent exe quits without the ipc being connected or an error. This is basically a (slightly) more immediate response than the timeout we already have.