allow mod types to be marked "non-essential"
what this means is that Vortex will allow using a deployment method even
if it isn't able to deploy the specified mod type.
Example: the game may read mods from a drive different from the gam files
.In that case hard link deployment would never be available since
enb/dinput mod types have to deploy to the game directory. By marking
enb/dinput as non-essential the game can be used with hardlink
deployment but Vortex will not be able to deploy enb/dinput mods