improved handling of permission protected mod directories
The node-permissions library can now deal with cases where the directory is not owned by admin (e.g. TrustedInstaller).
In addition the ensureDirWritable function will now acquire the necessary permissions recursively, so if the parent directory of the directory being created is write protected, _it's_ permissions are also changed to allow it.
Finally, vortex will no longer error out if the game binary isn't readable due to file permissions. This is again a problem with windows store games