allow bbcode dialogs to trigger internal callbacks
this is only currently used in a single location which is the test runner where it can be used to to have links that don't lead to external pages but open Vortex screens.
This has to be used very selectively to ensure we don't open up a security issue with bbcode supplied by websites.
Technically we can now provide any kind of context to the bbcode parser but the link tag is currently the only one that uses it