improvements/fixes to use of metadb information
a) There is now prioritization of the lookup results. If one of the results matches
the file name exactly, that result is preferred over all other results. Otherwise
the one providing the most information is preferred (this is a simplification, there is also prioritization on _which_ information is available)
b) the metadb information stored with a download is kept separate and is thus easier to ignore/overwrite with more reliable information (the metadb info can
currently not be considered reliable since the data on the nexus mods index is not curated)
c) this also includes fixes to the prioritization of attribute extractors (the functions that collect information about a mod during installation)