fixed: mod directories not removed on purge due to missing dir tags
Given that newly added files are deployed concurrently, the creation of
deep directory structures could potentially cause some mod folders to be
omitted when generating directory tags; these are used by Vortex to
keep track of which folders it has created and needs to remove upon mod
removal and/or purge events. We now implement a callback which will be
called upon each successful directory creation to create the directory
tags.