fixed: attempt to activate undiscovered game on startup
pretty rare combination of events where Vortex was closed while in the process of switching to a different game and then when Vortex got started again, the previous game was no longer discovered. Since Vortex, if interrupted during game switch, will revert to the previous profile, it would attempt to activate the that undiscovered game.