give a more "friendly" error message if runtask fails
we don't actually know why this would fail. The error code we got (4902) isn't even documented by microsoft which may indicate it gets generated by a foreign software or driver injected into the process.
The only way we will make progress here is if a knowledgeable user encounters this who could identify the setting/driver/software responsible for this.

fixes #6552