Supress report of ProcessCanceled errors on single mod deployment
ProcessCanceled error handling for single mod deployment is now in
line with how we cater for these types of errors during regular
deployment events. These types of errors are generally down to a fault
in the user's environment e.g. when the OS refuses to run a process
which requires elevation without providing an error message.

closes #6913