trim whitespace from staging directory
when installing a mod the name of the staging directory is derived from the archive name without the extension. If the name contained a space as the last character before the extension (foobar .zip) the staging folder would have a space as the last character.
Which is technically legal on NTFS at least but applications struggle to deal with directory names like that