fixed: modType mergeMods functors not used on mod removal
Previously we had two separate functions for the same logic which
was intended to resolve the subdir of a deployed mod. We were correctly
using the modtype mergeMods functor (when applicable) to resolve the
subdir, but we weren't doing so on removal; causing Vortex to leave
orphaned links behind as it was unable to find them + obviously raise
the external changes dialog as the mod's source would be removed, but
not the destination/links. Thankfully this should be a rare bug, only
seen in Star Wars: Fallen Order and the unreleased Blade and Sorcery
Load order enhancement.