allow users to select existing staging/download folders
That is: the settings dialog now allows users to select an existing, non-empty folder for the staging or download folder *if* it was tagged previously by Vortex.
If it was tagged by a different instance of Vortex there is a stern warning but the user can continue.
The main use case is that a user can, after deleting the Vortex appdata directory, set it up to use the previous folders, although the meta data is of course still gone