fixed: incomplete cleanup after installing nested installers
that is: if a zip contains a dragon age dazip (afaik the only concrete case where this could happen) the installer will install that dazip. If it was contained in a subdirectory, the temporary directory used for that was not cleaned up afterwards and on the next start would be picked up as a new mod.