added mechanism to catch react callback errors
react has a mechanism to catch exceptions during component rendering but with callbacks we're basically on our own, unless we have catches absolutely everywhere they will end up unhandled.
This now introduces a generic mechanism that can be used which will be caught and presented in the same way as rendering errors.
As a convenience these wrapped callbacks are also cached which frees the caller from doing that, although I'm not sure yet how this compares to React.useCallback in terms of performance and reliability.
This is not finalized and not exported to the api atm.