fixed: files lost trying to move staging/download folder onto itself
this would happen if the user tries to change their staging or download folder but the only difference is in the case of letters (a -> A). Since the file system is case insensitive, the directory is the same. This lead to cleanup code removing files under the assumption they had been copied, when they hadn't been.

fixes #7391