added api to mark text for translation
sometimes texts get stored during registration and wouldn't be updated before actually getting translated. To ensure the correct namespace gets used and it gets updated when the language changes this introduces a new class that can be stored instead of the string with the information required to translate it correctly later on.