validate application state in renderer
On startup the application state is transferred from the main process to renderer. There are bug reports that seem to indicate this parsing crashes for some users in native code and we have no way of figuring out whether the json data is just invalid in a weird way or it's running out of memory.
This change checksums the data before we try to parse it, if the crashes continue then it has to be oom.
We also now log the size of the buffer before parsing