don't offer to install mod for a game that isn't discovered
When a mod is installed that is intended for a game different than the one being managed, we show a dialog asking whether the user wants to install for the intended game or for the managed one.
Previously we only offered the "intended" game if that had been managed before, we didn't check whether it was actually discovered right now, which could lead to problems with various installers.