retry copyFile on EEXIST
the thinking is that we already assert the destination of the copy isn't the source and in all other cases the copy should overwrite the target, not fail.
The errors being reported were about mkdir failing which probably means the same dir was created asynchronously which is not a problem.