fixed: "give permission" failed if file doesn't exist
to elaborate: If creating a file fails because the directory it's being created in is missing write permission for the user, Vortex would try to acquire write access to that file (because that's what the EPERM error message is referring to) - which would fail because obviously we can't change permission on a file that doesn't exist.
The solution is to instead change permission on the directory.

This also fixes an error where fs errors weren't handled gracefully when copying a state backup

fixes #8166