Enhanced gamestore helper to be able to query if store is installed
Previously the gamestore helper did not provide the ability to verify
whether a gamestore is installed directly; we could indirectly verify
this by verifying if we're able to resolve the gamestore's installation
path using the getGameStorePath function (which was optional to
implement) this was completely unreliable when it came to verifying if
the xbox game store is installed.

Additionally the only reason why the getGameStorePath function was
optional in the first places was due to the xboxstore not being able to
provide an executable path like ALL other game stores. Rather than
keep this function optional for the benefit of a single gamestore;
store implementations are now able to define their own logic to determin
whether the store is installed correctly or not.

closes nexus-mods/vortex#8393