support for installers putting files into download directory
This is a fairly specific use case for bundled mods in collections.
Technically we could probably treat them as sub-installers but that would mean bundled mods get treated different and would appear different in the ui from all the other mods included in the collection.
The way it is now, bundled mods only differ in that the collection ships the archives for the mod.

This commit also fixes a few minor issues surrounding (fuzzy) version matching