better handling of multiple games managing same directory
In the case of total conversions or third-party game extensions for games that are already supported by Vortex you might have multiple extensions trying to manage the same directory.
This wasn't handled correctly previously and could cause a mess.

This commit contains two changes to remedy this:
a) deployment manifests now store which game extension they were created by. If during deployment a manifest is discovered from a different game, Vortex will now ask to purge it, instead of offering to change the staging folder (which is what happens if they are the same game because then Vortex assumes the user messed up their configuration)
b) a game extension can specify that it overrides another extension.
If those extension manage the same game directory, the one with the override specification takes precedence and the other gets disabled (it still appears but can't be managed anymore, with a message explaining why)
In case of total conversions that modify the directory of the original game, the extension should ensure that it only discoveres the game if the tc is actually installed.