fixed crash managing a game that uses the harmony injector on setup
The injector would use the activeGameId in its attempts to verify the
game's location. If the injector were to be called during the setup
stage of a game extension, the activeGameId would be set to the wrong
gameId or be completely undefined. This would cause the Injector to
complain about missing game assemblies.

Crash was being caused by poor handling of the missing assemblies error

fixes nexus-mods/vortex#8831