no longer ignore failure to load extensions
previously, if any extension (static, bundled or third-party) fails to load we would log the problem and continue. For dynamic extensions that error message was also visible on the Extensions page.
For third party extensions this makes sense because we don't want Vortex to break because of them - as far as possible - but continuing despite problems can cause data loss if the user restores a state backup because only those parts of the backup get restored that are actually known to Vortex at the time of the restore.
So if an extension fails to load and then the user restores a state backup, all data for that extension is lost.
Therefore now we will allow Vortex to crash if a static extension fails to load.

Further, we now no longer execute init calls from extensions that failed to initialize because previously we might have partially loaded extensions which could cause further issues with the extension trying to access state data that wasn't available because the registerReducer calls didn't get run.

Finally, all registerReducer calls were moved to the top of the respective init calls.
This isn't technically necessary after the above change but just to make sure that if it didn't work at least if anything from an extension gets run it's extending the state.