fixed: eventual download fail if disk can't keep up
apparently what happens is that the download is faster than the disk can write data to disk, as a result the buffer accumulates and eventually we fail to allocate memory to merge the buffers before writing.

The solution is: if merging buffers fails, we pause the download and write out the individual chunks before resuming the download.
In the future it may make sense to have a deterministic limit to how large the buffer may get instead of waiting for memory allocations to fail but for now this is a very very rare issue.

fixes #8322