fixed crash if uninstalling mod from history tab while re-installing mod
Previously we would not check if the mod is actually fully installed
when deciding whether to provide the user with the ability to uninstall
a mod through the history tab. So if a user would install a mod and then
uninstall it (decide he messed up) and went to re-install the mod again,
the older installation instance in history would still allow users to
attempt to uninstall the mod again even though the mod may have not
finished installing yet. Confused? so am I!

fixes nexus-mods/vortex#9057