performance improvements
a) after resolving a mod reference, if possible, we store the local id it was matched to. Next time the reference is to be resolved, Vortex will check the rule against that local mod first, skipping the remainging search process if it matches.
b) no longer trigger a refresh of the load order upon changing rules that have nothing to do with load order