fixed: correcting tags when mod already downloaded
To ensure Vortex correctly identifies mods downloaded/installed as a dependency, even if there are no identifiers we could base it on (e.g. downloads from foreign sites) we tag downloads installed this way. If we find and use an existing download we can't change/update a potentially existing tag because that might break the dependency that initially pulled it in, so instead we have to update the tag stored in the (new) dependency to match the existing file.