added "other" mod source, heuristic for fileid lookup
there were a few changes accidentally included in f2c204659a95db64189b8911eec0877c5aa449e7 that belong to this commit.
basically we're increasing the chance for Vortex to correctly figure out mod meta identication for a mod if it wasn't available during install.
The "other" mod source is there to distinguish mods that are explicitly not from any "repository" that would allow us to visit the corresponding website or check for updates - in contrast to the source being unset which means we don't know what the source is and might query the user