fixed "duplicate" download entries when replacing an existing download
Replacing an existing download will now replace the same archive file
name - previously an additional archive would be created with a '.1'
appended prior to the file extension.

Initially this was believed to be a simple visual issue but it could
potentially cause issues when replacing an archive as the old filename
would remain orphaned in the downloads folder without an application
state entry (the replace functionality will update the localPath
property) - upon Vortex restart, the orphaned downloads will get picked
up as locally imported downloads.

Imagine attempting to replace a broken/failed download, and then
attempting to install that faulty archive when Vortex is restarted.