improved error handling of missing/unknown downloads due to be installed
The 'start-install-download' event listener is used throughout the
application to install mod archives. Some event emitters are _not_
executed immediately, relying on user input to trigger the event. (e.g.
via the "Install All" action presented to the user through a notif after
a set of downloads have completed successfully)

If for any reason a download becomes unavailable (is removed from the
app state) before the user triggers the installation event, Vortex used
to simply report that it attempted to install an "unknown download".
Unfortunately no data is available at that stage to inform the user what
download failed.

This commit ensures that the error gets returned to the event emitter
via the callback functor, which _should_ ensure that more information
is available when the error is reported - unfortunately if the emitter
does _not_ provide a callback functor, the non descriptive "unknown
download" is still raised (better than nothing).

Finally, depending on how existing event emitter callbacks are handling
the DataInvalid error object - this change may produce new errors we
haven't seen before. I'm currently reviewing all existing emitters to
ensure nothing major occurs, but there a LOT of them in our bundled
extensions and core application (not to mention the community
extensions)