fixed mod files not being deployed due to redundant file overrides
This is part of the mod-dependency-manager commit
39df0b275ef2945a5d1969951f5f2492189e8f6a which intends to fix a use case
where the user defines a file override, and removes that file from the
staging folder. This used to leave the file override behind causing that
specific mod file to no longer be deployable (regardless of mod)

Any deletion action of a staging folder mod file detected by the
external changes functionality will now be propagated via an event to
assist the mod-dependency-manager extension to respond to it by checking
if any fileOverrides have been defined for that file, and if removal is
necessary.

fixes Nexus-Mods/Vortex#9671