install dependencies in phases
this does not include any way of assigning phases, just the low-level functionality.

Currently each phase gates the download and installation. It should be safe to start downloading for the next phase while the previous one is installing/deploying but the scheduling for that would be a mess because we'd still need to prioritize the downloads for the current phase.