modified discovery filepath assertion to improve game detection
This improves UX somewhat as the game thumbnail is highlighted in full
colour even if the game extension can't handle the game installation
for some reason. Any missing filepaths will raise an error notification
explaining _why_ the filepath may be missing which should hopefully
assist the user in debugging and solving the problem on his own.