initial support for sorting the game list
this also fixes an issue where we didn't fetch the list of games from nexus if the user wasn't logged in. In the past we did that through the api but since it's now just a publicly available json file there is no reason to not do it for all users.

Finally, the extension list may no include the game id/nexus domain for a game to help match third-party extension to the game they support.