fixed wrong game id assigned to downloads with >1 compatible extensions
Previously any new downloads would get assigned with the mod id of
compatible extensions as a priority instead of the id of the original
game extension.

e.g. all XCom2 mods would get assigned with the WOTC ext gameId instead,
which would cause the "No compatible game being managed" dialog to be
raised when installing the mod if the user did not have WOTC installed.